// apps/web/src/types/pokemon.ts
export type PokemonType
    = | 'Normal'
        | 'Fire'
        | 'Water'
        | 'Electric'
        | 'Grass'
        | 'Ice'
        | 'Fighting'
        | 'Poison'
        | 'Ground'
        | 'Flying'
        | 'Psychic'
        | 'Bug'
        | 'Rock'
        | 'Ghost'
        | 'Dragon'
        | 'Dark'
        | 'Steel'
        | 'Fairy';

export interface Pokemon {
    id: number;
    name: string;
    types: PokemonType[];
    sprite: string; // URL to a 64x64 sprite
    description: string;
}